Use ReSpec's generated heading level
ReSpec ignores the provided heading level (h1, h2) and assigns a
heading level based on the section nesting.

As a result, it is convention to use h2 for all headings. This allows
section nesting to be changed on the fly without worrying about
updating the headings.

This commit changes all headings to h2.