Fix ReSpec section titles
ReSpec requires that the first element of a section be a heading. It
uses that heading for the title, which then becomes an entry in the
index.

Currently, the first element is the anchor which preserves old
fragments. This breaks ReSpec's expectations.

This commit uses ids instead of anchors and re-establishes ReSpec's
expectations.

Closes #74