Fix respec anchor targets (#77)
* Use HTTPS for ReSpec fetch

Currently, ReSpec is fetched via HTTP.
If the spec is hosted via HTTPS, the HTTP ReSpec request will be denied
in modern browser settings. This results in ReSpec styling not being
applied.

This commit fetches ReSpec via HTTPS.

* Fix ReSpec section titles

ReSpec requires that the first element of a section be a heading. It
uses that heading for the title, which then becomes an entry in the
index.

Currently, the first element is the anchor which preserves old
fragments. This breaks ReSpec's expectations.

This commit uses ids instead of anchors and re-establishes ReSpec's
expectations.

Closes #74