Fix broken table fragments (#100)
Tables were using the old fragment style <a name="frag_name"></a>.
ReSpec correctly errored on these.

ReSpec has a special way to handle figures. But it doesn't yet have a
special way to handle tables. There is discussion about this here:
https://github.com/w3c/respec/issues/4004

This commit will leave the tables in place for now (not using any
special ReSpec handling) but corrects the fragments. It also adds a
comment to preserve fragments. This fixes several ReSpec errors.

Closes #99