Use ReSpec for internal links (1 of N)
This patch is the first in a series which will use ReSpec's formatting
for internal links.

It is broken into pieces to make review manageable.

ReSpec will format non-definition internal links with the auto-generated
section number. Definition internal links do not receive any special
styling. But using them is forward compatible if ReSpec later decides to
apply formatting.

This is part of #122 and #123.