Merge iCCN chunk into iCCP chunk
Previously, we considered the iCCP chunk should be for ICC v2 profiles
only. This is what prior editions of the spec made explicit. So a new
chunk (iCCN) would allow ICC v4 profiles.

However, it was decided that the existing iCCP chunk can safely contain
ICC v4 profiles and a new chunk is not needed.

This commit moves the comments about the iCCN chunk into the iCCP chunk,
allowing the iCCP chunk to handle ICC v4 profiles.