Update table styling
ReSpec was recently updated to support table tracking the same way it
supports figures. This landed in https://github.com/w3c/respec/pull/4200

ReSpec already supported class="simple" on tables to gain styling.

This commit adds both of these to all tables in the PNG spec.
This also means adding captions and headings to tables which were
missing it.

Closes #138