Replace '<span class="chunk">' with links.
The 'chunk' class is unused. However, keeping it around could be useful
for later.

Rather than keep it as a span, it should be an anchor link.

This commit replaces '<span class="chunk">' with anchor links.

Note: '<span class="chunk">' is left intact when defining the chunk
itself. For example, the pHYs chunk doesn't link to itself.