Don't use <tt> tag
The W3C HTML Validator requires that the obsolete <tt> tag is not used.
It recommends some alternatives. For our application, the <code> and
<var> tags are useful.

When a <var> is clicked, it creates an additional search, highlighting
all matching <var>s. A user can click multiple different <var>s to
highlight multiple <var>s. For this reason, variable text should mean
the same variable in all uses. For example, <var>x</var> should always
refer to the same "x".

This commit removes the <tt> tag, replacing it with <code> and <var>
where appropriate.