Fix ReSpec error
There is an <a> element that doesn't link anyway. It is given an id to
be a link destination. ReSpec does not understand the intent of this
empty, no-destination anchor element. ReSpec errors on this strange
situation.

This commit removes the erroring <a> tag and maintains the id on a more
appropriate tag.