Only error on new GitHub workflow issues
Currently, the GitHub workflow (which runs on every pull request) will
report errors if the existing files contain errors. This is unfortunate
because people who create new pull requests will see an error when they
did not cause it.

This commit will instead only error if the pull request introduces a new
error.