Move editors to editor section
Two new editors were added to the authors section.
I believe the authors section is intended for the original authors of 1.0.
When someone updates the spec, I believe they should be listed as an
editor.

This commit moves two people from the authors section to the editors
section to reflect this.