Correct JSON commas (#468)
In commit 1d54517, some JSON objects were moved around. The missing
trailing comma typical to JSON was not added back for the non-trailing
case. This caused a parsing error.

This commit corrects the commas from that previous commit.

It also removes a trailing comma that was apparently previously allowed.
This matches the style of the rest of the JSON. Although, perhaps we
should opt into trailing commas to make life easier.